requirements.txt
streamlit
google-generativeai
python-dotenv
fastapi
uvicorn
python-multipart
pydantic